requirements.txt
streamlit==1.41.1
easyocr==1.7.2
torch
torchvision
opencv-python-headless
numpy<2.0
Pillow==11.0.0
groq==0.13.1
python-dotenv==1.0.1
